// src/lib/shyft/models/lpInfo.model.ts
export type LpInfo = {
  _updatedAt: string;
  poolId: string;
  amountWaveRatio: number;
  baseDecimal: number;
  baseLotSize: number;
  baseMint: string;
  baseNeedTakePnl: number;
  baseTotalPnl: number;
  baseVault: string;
  depth: number;
  lpMint: string;
  lpReserve: string;
  lpVault: string;
  marketId: string;
  marketProgramId: string;
  maxOrder: number;
  maxPriceMultiplier: number;
  minPriceMultiplier: number;
  minSeparateDenominator: number;
  minSeparateNumerator: number;
  minSize: number;
  nonce: number;
  openOrders: string;
  orderbookToInitTime: number;
  owner: string;
  pnlDenominator: number;
  pnlNumerator: number;
  poolOpenTime: number;
  punishCoinAmount: number;
  punishPcAmount: number;
  quoteDecimal: number;
  quoteLotSize: number;
  quoteMint: string;
  quoteNeedTakePnl: number;
  quoteTotalPnl: number;
  quoteVault: string;
  state: number;
  status: number;
  swapBase2QuoteFee: number;
  swapBaseInAmount: number;
  swapBaseOutAmount: number;
  swapFeeDenominator: number;
  swapFeeNumerator: number;
  swapQuote2BaseFee: number;
  swapQuoteInAmount: number;
  swapQuoteOutAmount: number;
  systemDecimalValue: number;
  targetOrders: string;
  tradeFeeDenominator: number;
  tradeFeeNumerator: number;
  volMaxCutRatio: number;
  withdrawQueue: string;
  pubkey: string;
};